autocorr/cov fixes

+++ Environ/TSA/sampleACF.hpp
@@ -5,6 +5,6 @@ using namespace std;
 
 double sm(vector<bar> v);
 
-long s_autocovf(vector<bar> v, int h);
+double s_autocovf(vector<bar> v, int h);
 
-long sample_acf(vector<bar> v, int h);
+double sample_acf(vector<bar> v, int h);
